backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
instagrapi
pillow
google-generativeai
APScheduler
